Add useTodoState custom hook to refactor
