+++ src/components/TodoApp.js
@@ -2,6 +2,8 @@
 import React from "react";
 import { jsx } from "@emotion/core";
 
+import useTodoState from "../hooks/useTodoState";
+
 import TodoList from "./TodoList";
 import TodoForm from "./TodoForm";
 
@@ -9,7 +11,6 @@ import Paper from "@material-ui/core/Paper";
 import AppBar from "@material-ui/core/AppBar";
 import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
-import uuid from "uuid/v4";
 
 const paperHeight = {
   padding: "0px",
@@ -20,41 +21,18 @@ const paperHeight = {
 
 function TodoApp() {
   const initialTodos = JSON.parse(window.localStorage.getItem("todos") || "[]");
+  const { todos, addTodo, removeTodo, toggleTodo, editTodo } = useTodoState(
+    initialTodos
+  );
   // const initialTodos = [
   //   { id: 1, task: "Learn React-Redux", completed: false },
   //   { id: 2, task: "Wash Car", completed: false },
   //   { id: 3, task: "Deploy Project", completed: true }
   // ];
 
-  const [todos, setTodos] = React.useState(initialTodos);
-
   React.useEffect(() => {
     window.localStorage.setItem("todos", JSON.stringify(todos));
   }, [todos]);
-
-  function addTodo(newTodo) {
-    setTodos([...todos, { id: uuid(), task: newTodo, completed: false }]);
-    console.log(uuid());
-  }
-
-  function removeTodo(todoId) {
-    const updatedTodos = todos.filter(todo => todo.id !== todoId);
-    setTodos(updatedTodos);
-  }
-
-  function toggleTodo(todoId) {
-    const updatedTodos = todos.map(todo =>
-      todo.id === todoId ? { ...todo, completed: !todo.completed } : todo
-    );
-    setTodos(updatedTodos);
-  }
-
-  function editTodo(todoId, newTask) {
-    const updatedTodos = todos.map(todo =>
-      todo.id === todoId ? { ...todo, task: newTask } : todo
-    );
-    setTodos(updatedTodos);
-  }
 
   return (
     <Paper style={paperHeight} elevation={0}>
